Cleanup the use of Double primitive wrapper.
Cleanup the use of Double primitive wrapper. As of Java 9, the VM can do a better job of handling the conversion from a Double primitive to a Double Object or vice versa.

This is an extension of the work done in #2498.

No tests have been touched as those expressly need to manage the conversion.

A couple of places I also cleaned up the Float.

+++ java/j2ee.persistence/test/unit/src/org/netbeans/test/stub/api/Stub.java
@@ -157,9 +157,9 @@ public final class Stub {
                 } else if (retClass == Long.TYPE) {
                     return 0L;
                 } else if (retClass == Float.TYPE) {
-                    return Float.valueOf(0);
+                    return 0F;
                 } else if (retClass == Double.TYPE) {
-                    return 0.0;
+                    return 0D;
                 } else if (retClass == Character.TYPE) {
                     return '\0';
                 } else if (retClass == Boolean.TYPE) {
